backend/requirements.txt
asgiref==3.9.1
attrs==25.3.0
bcrypt==4.3.0
cffi==1.17.1
coverage==7.10.5
cryptography==45.0.6
Django==5.2.5
django-cors-headers==4.7.0
django-debug-toolbar==6.0.0
django-redis==6.0.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-spectacular==0.28.0
Faker==37.5.3
gunicorn==23.0.0
inflection==0.5.1
iniconfig==2.1.0
jsonschema==4.25.1
jsonschema-specifications==2025.4.1
packaging==25.0
pluggy==1.6.0
psycopg2-binary==2.9.10
pycparser==2.22
Pygments==2.19.2
PyJWT==2.10.1
pytest==8.4.1
pytest-django==4.11.1
python-decouple==3.8
PyYAML==6.0.2
redis==6.4.0
referencing==0.36.2
rpds-py==0.27.0
sqlparse==0.5.3
typing_extensions==4.14.1
tzdata==2025.2
uritemplate==4.2.0
